% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpkg.R
\name{findpkg}
\alias{findpkg}
\title{Find package by query words}
\usage{
findpkg(data, query)
}
\arguments{
\item{data}{results of functin loadData()}

\item{query}{one or more query words}
}
\value{
relative packages
}
\description{
Find package from package description field.
}
\examples{
\donttest{
d <- loadData()
findpkg(d,c('table','one'))
}
}
