% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2package.R
\name{fun2package}
\alias{fun2package}
\title{Find package by a keyword in function}
\usage{
fun2package(data = NULL, funkey, exact = FALSE, repo = getOption("repos"))
}
\arguments{
\item{data}{NULL or result of loadData() function.}

\item{funkey}{one character}

\item{exact}{logical, whether to match exactly}

\item{repo}{repository}
}
\value{
a datafram contains package and function
}
\description{
Find package by a keyword in function
}
\examples{
\donttest{
fun2package(funkey='nomogram')
}

}
