% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.most.R
\name{imports.most}
\alias{imports.most}
\title{Mostly imported packages}
\usage{
imports.most(imports, n = 10, year = FALSE)
}
\arguments{
\item{imports}{results of function imports()}

\item{n}{the most frequency number}

\item{year}{logical, default is FALSE. Whether to include years}
}
\value{
a data.frame contains package, frequency or year
}
\description{
Mostly imported packages
}
\examples{
\donttest{
    d <- loadData()
    i <- imports(d)
    imt <- imports.most(imports = i,20)

    library(ggplot2)
    ggplot(imt,aes(Imports,Freq))+
        geom_col()+
        coord_flip()+
        theme(axis.text = element_text(size=16))


    imt <- imports.most(i,10,T)
    imt <- imt[imt$year >= 2011 & imt$year <= 2020,]
    # the latest 10 years
    library(ggplot2)
    library(tidytext)

    ggplot(imt,aes(reorder_within(Imports,Freq,year),Freq))+
        geom_col()+
        scale_x_reordered() +
        facet_wrap(~year,scales="free")+
        coord_flip()+
        theme(
            axis.text = element_text(size=16),
            strip.text.x = element_text(size = 18,
                                        colour = "red")
        )+
        xlab(NULL)+ylab(NULL)
}
}
