% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.plot.R
\name{imports.plot}
\alias{imports.plot}
\title{Plot for imports frequency data}
\usage{
imports.plot(imports.most)
}
\arguments{
\item{imports.most}{results of function imports.most()}
}
\value{
a ggplot2 style plot
}
\description{
Plot for imports frequency data
}
