% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{imports}
\alias{imports}
\title{Imports packages of one package}
\usage{
imports(data)
}
\arguments{
\item{data}{data of function loadData()}
}
\value{
a dataframe contains Package, Imports and modified time
}
\description{
Imports packages of one package
}
\examples{
\donttest{
d <- loadData()
i <- imports(d)
head(i)
}
}
