% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.sankey.R
\name{imports.sankey}
\alias{imports.sankey}
\title{Plot sankey picture}
\usage{
imports.sankey(imports.most)
}
\arguments{
\item{imports.most}{results of function imports.most()}
}
\value{
a sankey plot
}
\description{
Plot sankey picture
}
